+++ gql_http_link/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- ability to configure headers on link
+- read headers from `Context`
+- pass HTTP status code in `Response.context`
+- improved exception handling
+
 ## 0.1.2
 
 - upgrade `package:gql` to v0.9.0